requirements.txt
PySerial
boltons